authentication added

+++ app/src/main/java/com/example/routing/Feedback.java
@@ -7,7 +7,7 @@ public class Feedback {
     public String dest;
     public String destLatLng;
     public String waypoints;
-    public int Q1;
+    public float Q1;
     public boolean Q2, Q3, Q4, Q5;
 
     public Feedback(){
@@ -18,7 +18,7 @@ public class Feedback {
         this.Q5 = false;
     }
 
-    public Feedback(long timestamp, String or, String originLatLng, String de, String deLatLng, String waypoints, int Q1, boolean Q2, boolean Q3, boolean Q4, boolean Q5){
+    public Feedback(long timestamp, String or, String originLatLng, String de, String deLatLng, String waypoints, float Q1, boolean Q2, boolean Q3, boolean Q4, boolean Q5){
         this.timestamp = timestamp;
         this.origin = or;
         this.originLatLng = originLatLng;
